+++ src/events/RegisterCustomFieldAdaptersEvent.php
@@ -1,0 +1,12 @@
+<?php
+namespace verbb\formie\events;
+
+use yii\base\Event;
+
+class RegisterCustomFieldAdaptersEvent extends Event
+{
+    // Properties
+    // =========================================================================
+
+    public array $adapters = [];
+}